feat: nested stack as per subnet group
