+++ src/index.ts
@@ -1,1 +1,2 @@
 export * from './constructs/network';
+export * from './constructs/vpc-endpoint-service';
